+++ prisma/migrations/20260909125951_add_analytics_field_derived_columns/migration.sql
@@ -1,0 +1,16 @@
+-- AlterTable
+ALTER TABLE "AnalyticsEvent" ADD COLUMN     "browser" TEXT,
+ADD COLUMN     "operatingSystem" TEXT,
+ADD COLUMN     "referrerDomain" TEXT;
+
+-- CreateIndex
+CREATE INDEX "AnalyticsEvent_country_idx" ON "AnalyticsEvent"("country");
+
+-- CreateIndex
+CREATE INDEX "AnalyticsEvent_referrerDomain_idx" ON "AnalyticsEvent"("referrerDomain");
+
+-- CreateIndex
+CREATE INDEX "AnalyticsEvent_operatingSystem_idx" ON "AnalyticsEvent"("operatingSystem");
+
+-- CreateIndex
+CREATE INDEX "AnalyticsEvent_browser_idx" ON "AnalyticsEvent"("browser");